requirements.txt
python
python-lxml 
python-psycopg2
python-boto
unpaper
tesseract-ocr
curl
poppler-utils
parallel
runit
